Refactor cgroup metrics to server.agent
