+++ mantis-server/mantis-server-agent/src/main/java/io/mantisrx/server/agent/cgroups/CgroupImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.mantisrx.server.worker.metrics.cgroups;
+package io.mantisrx.server.agent.cgroups;
 
 import io.mantisrx.shaded.org.apache.curator.shaded.com.google.common.base.Preconditions;
 import io.vavr.Tuple2;
