+++ mantis-server/mantis-server-agent/src/main/java/io/mantisrx/server/agent/cgroups/NetworkSubsystemProcess.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.mantisrx.server.worker.metrics.cgroups;
+package io.mantisrx.server.agent.cgroups;
 
 import io.mantisrx.server.master.client.config.Usage.UsageBuilder;
 import java.io.IOException;
