+++ mantis-server/mantis-server-agent/src/main/java/io/mantisrx/server/agent/cgroups/SubsystemProcess.java
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package io.mantisrx.server.worker.metrics.cgroups;
+package io.mantisrx.server.agent.cgroups;
 
+import io.mantisrx.server.master.client.config.Usage;
 import java.io.IOException;
-import java.util.Map;
 
-interface Cgroup {
-    Long getMetric(String subsystem, String metricName) throws IOException;
-    Map<String, Long> getStats(String subsystem, String stat) throws IOException;
+interface SubsystemProcess {
+    void getUsage(Usage.UsageBuilder usageBuilder) throws IOException;
 }
